# contracts/marketplace-customer.openapi.yaml
openapi: 3.1.0
info:
  title: Tellbook Marketplace Customer Data API
  version: 1.0.0
servers:
  - url: /v1/marketplace
security:
  - marketplaceSession: []
paths:
...
  /auth/code:
    post:
      operationId: startMarketplaceAuthCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeStartInput' }
      responses:
        '202': { $ref: '#/components/responses/AuthChallengeAccepted' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
        '429': { $ref: '#/components/responses/AuthRateLimited' }
  /auth/code/{challengeID}:
    get:
      operationId: getMarketplaceAuthCodeStatus
      security: []
      parameters:
        - name: challengeID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/AuthChallengeReady' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
  /auth/code/resend:
    post:
      operationId: resendMarketplaceAuthCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeResendInput' }
      responses:
        '202': { $ref: '#/components/responses/AuthChallengeAccepted' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
        '429': { $ref: '#/components/responses/AuthRateLimited' }
  /auth/verify:
    post:
      operationId: verifyMarketplaceAuthCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeVerifyInput' }
      responses:
        '200':
          description: Marketplace session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceCodeSessionResponse' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
  /auth/password:
    post:
      operationId: marketplacePasswordSignIn
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordSignInInput' }
      responses:
        '200':
          description: Marketplace session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password/reset/code:
    post:
      operationId: startMarketplacePasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordResetStartInput' }
      responses:
        '202': { $ref: '#/components/responses/AuthChallengeAccepted' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
        '429': { $ref: '#/components/responses/AuthRateLimited' }
  /auth/password/reset/verify:
    post:
      operationId: verifyMarketplacePasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordResetVerifyInput' }
      responses:
        '200':
          description: Single-use reset grant.
          content:
            application/json:
              schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordResetVerifyResponse' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
  /auth/password/reset:
    post:
      operationId: completeMarketplacePasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordResetCompleteInput' }
      responses:
        '200':
          description: Password updated and a replacement session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSessionResponse' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
  /auth/session:
    get:
      operationId: getMarketplaceSession
      security: []
      responses:
        '200':
          description: Current marketplace session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: refreshMarketplaceSession
      security: []
      responses:
        '200':
          description: Current marketplace session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /me/identities/code:
    post:
      operationId: startMarketplaceIdentityLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeStartInput' }
      responses:
        '202': { $ref: '#/components/responses/AuthChallengeAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/identities/code/{challengeID}:
    get:
      operationId: getMarketplaceIdentityLinkStatus
      parameters:
        - name: challengeID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/AuthChallengeReady' }
        '400': { $ref: '#/components/responses/BadAuthRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/identities/code/resend:
    post:
      operationId: resendMarketplaceIdentityLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeResendInput' }
      responses:
        '202': { $ref: '#/components/responses/AuthChallengeAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/identities/verify:
    post:
      operationId: verifyMarketplaceIdentityLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/CodeVerifyInput' }
      responses:
        '200':
          description: Identity linked to the current customer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplaceIdentityLinkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/BadAuthRequest' }
  /me/password:
    patch:
      operationId: updateMarketplacePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './auth.schemas.yaml#/components/schemas/PasswordUpdateInput' }
      responses:
        '200':
          description: Password updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketplacePasswordResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    marketplaceSession:
      type: apiKey
      in: cookie
      name: tellbook_marketplace_session
...
  responses:
    AuthChallengeAccepted:
      description: Challenge accepted without disclosing account existence.
      content:
        application/json:
          schema: { $ref: './auth.schemas.yaml#/components/schemas/ChallengeResponse' }
    AuthChallengeReady:
      description: Current delivery and verification state for the challenge.
      content:
        application/json:
          schema: { $ref: './auth.schemas.yaml#/components/schemas/ChallengeResponse' }
    BadAuthRequest:
      description: Invalid authentication request.
      content:
        application/json:
          schema: { $ref: './auth.schemas.yaml#/components/schemas/Error' }
    AuthRateLimited:
      description: Authentication request limit reached.
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content:
        application/json:
          schema: { $ref: './auth.schemas.yaml#/components/schemas/Error' }
    Unauthorized:
      description: Marketplace customer session is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    MarketplaceCustomer:
      type: object
      additionalProperties: false
      required: [id, full_name, has_password, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, pattern: '^\\+[1-9][0-9]{9,14}$' }
        email_verified_at: { type: string, format: date-time }
        phone_verified_at: { type: string, format: date-time }
        birthday: { type: string, format: date }
        has_password: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MarketplaceSessionResponse:
      type: object
      additionalProperties: false
      required: [customer]
      properties:
        customer: { $ref: '#/components/schemas/MarketplaceCustomer' }
    MarketplaceCodeSessionResponse:
      type: object
      additionalProperties: false
      required: [customer, is_new_account, onboarding_required]
      properties:
        customer: { $ref: '#/components/schemas/MarketplaceCustomer' }
        is_new_account: { type: boolean }
        onboarding_required: { type: boolean }
...
    MarketplacePasswordResponse:
      type: object
      additionalProperties: false
      required: [message, customer, reauthentication_required]
      properties:
        message: { type: string }
        customer: { $ref: '#/components/schemas/MarketplaceCustomer' }
        reauthentication_required: { type: boolean, const: true }
    MarketplaceIdentityLinkResponse:
      type: object
      additionalProperties: false
      required: [customer, reauthentication_required]
      properties:
        customer: { $ref: '#/components/schemas/MarketplaceCustomer' }
        reauthentication_required: { type: boolean, const: true }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...